fix: render qwik city slots in layouts
