+++ packages/qwik/README.md
@@ -17,9 +17,9 @@ npm install @ssgoi/qwik
 Qwik serializes component state for resumability, while SSGOI transition configs contain functions. Pass the config as a QRL factory so it is created in the browser when the observer starts.
 
 ```tsx
-import { $, Slot, component$ } from "@builder.io/qwik";
+import { $, Slot, component$, useSignal } from "@builder.io/qwik";
 import { Link, useLocation } from "@builder.io/qwik-city";
-import { Ssgoi } from "@ssgoi/qwik";
+import { useSsgoi } from "@ssgoi/qwik";
 import { drill, zoom } from "@ssgoi/qwik/view-transitions";
 
 const ssgoiConfig$ = $(() => ({
@@ -32,12 +32,16 @@ const ssgoiConfig$ = $(() => ({
 
 export default component$(() => {
   const location = useLocation();
+  const ssgoiRoot = useSignal<HTMLElement>();
+
+  useSsgoi(ssgoiRoot, { config$: ssgoiConfig$ });
 
   return (
-    <main class="relative z-0 h-dvh overflow-y-auto overflow-x-clip">
-      <Ssgoi config$={ssgoiConfig$}>
-        <Slot />
-      </Ssgoi>
+    <main
+      ref={ssgoiRoot}
+      class="relative z-0 h-dvh overflow-y-auto overflow-x-clip"
+    >
+      <Slot />
       <Link href="/posts/">Posts</Link>
       <p>{location.url.pathname}</p>
     </main>
@@ -58,3 +62,7 @@ export default component$(() => {
 ```
 
 Import view-level factories from `@ssgoi/qwik/view-transitions`.
+
+`Ssgoi` is also exported for wrapping concrete children. In Qwik City layouts,
+prefer `useSsgoi` because forwarding a route `<Slot />` through another
+component prevents Qwik City from projecting routed content into the DOM.
