+++ packages/qwik/src/lib/index.ts
@@ -1,4 +1,5 @@
 export * from "./types";
+export { useSsgoi, type UseSsgoiOptions } from "./use-ssgoi";
 export { Ssgoi } from "./ssgoi";
 export { SsgoiTransition } from "./ssgoi-transition";
 export * from "@ssgoi/core";
